v0.3.3; Added currentLanguageMatch and scripts
Added a function to get the current editor language, and then compare it against a list of options (with the option to coalesce certain language names to other values).

Also added scripts in package.json for updating the version of the package (and not updating if linting or anything else fails).