v0.3.4; updated type for currentLanguageMatch
Changed currentLanguageMatch to return string | null instead of string | null | any.

The any result was since the coalesce parameter allowed a return value of any type, but it has been restricted to returning only strings.

This restriction makes it so the user will not have to narrow the type of the result from any to string, since the intention was for coalesce to return strings 99% of the time — that 1% is now not allowed, but the pros outweigh the cons.