v0.3.17; Change previewmarkdownFile to return bool
Changed previewMarkdown file to capture any errors with previewing the markdown file, and return a boolean representing if the "markdown.showPreview" command was ran successfully or not.

This allows the user the option to check if the preview was unsucessful, and to react and/or fallback to something like just opening the markdown source instead.