v0.3.26; updated temporary messages
Updated function for temporary messages to fix a not-bug bug where the message progress counter would not at all match where it should be if the tick-rate was millisecond-precise.
This was because the timing-related functions within Node JS could be notably inaccurate (in terms of ms), and so a library "nanotimer" was imported to use much more precise timer functions to allow for a more accurate progress bar for temporary messages with precise millisecond ticks.

Also fixed some other accidental bugs and small errors, and updated doc comments for small changes to functions.